IR-gen: use more generic x86-64 level rather than native
This allows for better optimization potential when building on one target, and running on
another with similar CPU capabilities.

When I compiled on Icelake and ran on Genoa, I had a huge performance regression (~2x), and
similarly when compiling in the opposite way. This resolves that. Initial IR should be as
generic as possible while retaining user needs.